Migrate UTF8 decoding benchmarks to Benchmark.DotNet
